+++ lib/neumorphic_package_by_serge_software/neumorphic_card.dart
@@ -53,7 +53,7 @@ class NeumorphicContainer extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final color = decoration.color ?? Theme.of(context).colorScheme.background;
+    final color = decoration.color ?? Theme.of(context).colorScheme.surface;
     final emboss = curveType == CurveType.emboss;
 
     Color colorValue = color;